+++ ASCOfficeDocxFile2/BinReader/DocumentWriter.h
@@ -33,7 +33,7 @@
 #define DOCUMENT_WRITER
 
 #include "../../XlsxSerializerCom/Common/Common.h"
-#include "../BinWriter/BinReaderWriterDefines.h";
+#include "../BinWriter/BinReaderWriterDefines.h"
 
 namespace Writers
 {
@@ -78,7 +78,7 @@ namespace Writers
 			if(BinDocxRW::g_nCurFormatVersion < 5)
 			{
 				bool bTitlePage = false;
-				for(int i = 0, length = m_oHeaderFooterWriter.m_aHeaders.size(); i < length; ++i)
+				for(size_t i = 0, length = m_oHeaderFooterWriter.m_aHeaders.size(); i < length; ++i)
 				{
 					HdrFtrItem* pHeader = m_oHeaderFooterWriter.m_aHeaders[i];
                     if(false == pHeader->rId.empty())
@@ -98,7 +98,7 @@ namespace Writers
 						}
 					}
 				}
-				for(int i = 0, length = m_oHeaderFooterWriter.m_aFooters.size(); i < length; ++i)
+				for(size_t i = 0, length = m_oHeaderFooterWriter.m_aFooters.size(); i < length; ++i)
 				{
 					HdrFtrItem* pFooter = m_oHeaderFooterWriter.m_aFooters[i];
                     if(false == pFooter->rId.empty())
