+++ Common/DocxFormat/Source/XlsxFormat/Comments/Comments.h
@@ -93,7 +93,7 @@ namespace OOX
 			virtual void toXML(NSStringUtils::CStringBuilder& writer) const
 			{
 				writer.WriteString(L"<authors>");
-				for(unsigned int i = 0, length = m_arrItems.size(); i < length; ++i)
+				for(size_t i = 0, length = m_arrItems.size(); i < length; ++i)
 				{
 					writer.WriteString(L"<author>");
 					writer.WriteEncodeXmlString(*m_arrItems[i]);
@@ -217,7 +217,7 @@ namespace OOX
 			virtual void toXML(NSStringUtils::CStringBuilder& writer) const
 			{
 				writer.WriteString(L"<commentList>");
-				for(unsigned int i = 0, length = m_arrItems.size(); i < length; ++i)
+				for(size_t i = 0, length = m_arrItems.size(); i < length; ++i)
 				{
 					m_arrItems[i]->toXML(writer);
 				}
