+++ Common/DocxFormat/Source/XlsxFormat/Workbook/BookViews.h
@@ -133,18 +133,20 @@ namespace OOX
 			}
 
 		public:
-            virtual std::wstring      toXML() const
+            virtual std::wstring toXML() const
 			{
 				return _T("");
 			}
 			virtual void toXML(NSStringUtils::CStringBuilder& writer) const
 			{
 				writer.WriteString(_T("<bookViews>"));
-				for(unsigned int i = 0, length = m_arrItems.size(); i < length; ++i)
+				
+				for(size_t i = 0, length = m_arrItems.size(); i < length; ++i)
 					m_arrItems[i]->toXML(writer);
+				
 				writer.WriteString(_T("</bookViews>"));
 			}
-			virtual void         fromXML(XmlUtils::CXmlLiteReader& oReader)
+			virtual void fromXML(XmlUtils::CXmlLiteReader& oReader)
 			{
 				ReadAttributes( oReader );
